(feat) adds structure refactor - fixes promise error - removes old index.js
