+++ container_logs/nginx/access.log
@@ -1,0 +1,3 @@
+172.19.0.1 - - [22/Apr/2025:17:16:48 +0000] "GET /doc HTTP/1.1" 404 142 "-" "Mozilla/5.0 (X11; Ubuntu; Linux x86_64; rv:137.0) Gecko/20100101 Firefox/137.0" "-"
+172.19.0.1 - - [22/Apr/2025:17:25:40 +0000] "GET /doc/ HTTP/1.1" 200 749 "-" "Mozilla/5.0 (X11; Ubuntu; Linux x86_64; rv:137.0) Gecko/20100101 Firefox/137.0" "-"
+172.19.0.1 - - [22/Apr/2025:17:25:40 +0000] "GET /favicon.ico HTTP/1.1" 404 150 "http://localhost/doc/" "Mozilla/5.0 (X11; Ubuntu; Linux x86_64; rv:137.0) Gecko/20100101 Firefox/137.0" "-"
