+++ package.json
@@ -1,12 +1,27 @@
 {
-  "name": "wt1_cookstore",
-  "version": "1.0.0",
-  "description": "An app to store recipes",
-  "main": "index.js",
+  "name": "cookstore-next",
+  "version": "0.1.0",
+  "private": true,
   "scripts": {
-    "test": "echo 'Error: no test specified'; false",
-    "start": "node index.js"
+    "dev": "next dev --turbopack",
+    "build": "next build",
+    "start": "next start",
+    "lint": "next lint"
   },
-  "author": "",
-  "license": "ISC"
+  "dependencies": {
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "next": "15.3.2"
+  },
+  "devDependencies": {
+    "typescript": "^5",
+    "@types/node": "^20",
+    "@types/react": "^19",
+    "@types/react-dom": "^19",
+    "@tailwindcss/postcss": "^4",
+    "tailwindcss": "^4",
+    "eslint": "^9",
+    "eslint-config-next": "15.3.2",
+    "@eslint/eslintrc": "^3"
+  }
 }
